zfs tank exception and updated safe_delete

+++ drakrun/drakrun/util.py
@@ -122,8 +122,11 @@ def get_xen_commandline(parsed_xl_info):
 
 def safe_delete(file_path) -> bool:
     try:
-        os.remove(file_path)
-        logging.info(f"Deleted {file_path}")
+        if os.path.exists(file_path):
+            os.remove(file_path)
+            logging.info(f"Deleted {file_path}")
+        else:
+            logging.info(f"Already deleted {file_path}")
         return True
     except OSError as e:
         logging.warning(f"{e.filename}: {e.strerror}")
